[App Manifest + Docs] Tutorial Doc Update for New Manifest, Associated Fixes (#525)
Update the tutorial doc based on the changes from
https://github.com/grafana/grafana-app-sdk/pull/483 and
https://github.com/grafana/grafana-app-sdk/pull/491. Running through the
tutorial again ran into a few bugs that this PR also resolves:
* The template kind used `permissions` instead of `extraPermissions`
* The default plugin name uses the full group, which contains `.` and is
invalid. Fixed to use only the appName portion
* The updated kind in `cue.def` still had a reference to `apiResource`
in the version map for inheriting validation/mutation, fixed to not use
apiResource
* Prometheus metrics for watchers cannot be exposed in the `simple.App`
design as they can for `simple.Operator`, updated
`operator.OpinionatedWatcher` to expose metrics from wrapped watchers,
and a method to `simple.App` to allow exposing arbitrary prometheus
collectors to the runner.

The tutorial front-end has also been updated to use the grafana API
server, rather than the plugin back-end. The plugin back-end will be
phased out of the tutorial entirely in the near future.