Upgrade Go to 1.23.4 and linter to 1.62.2 (#535)
### What

* Upgrade to Go 1.23.4
* Remove Go version check from Makefile and make all default target
* Upgrade golangci-lint to 1.62.2 and fix issues

### Why

To make sure we are running the latest toolchain and that the code is
fixed accordingly. The changes to the Make target were done because Go
command is smart enough to download and install the version specified in
`go.mod` (assuming `GOOTOOLCHAIN=auto`) so the check is superfluous.
Moving it below `all` also makes it so that running `make` runs `make
all` again.

---------

Signed-off-by: Igor Suleymanov <igor.suleymanov@grafana.com>