[codegen] Generate exported constants for group and version. (#563)
Generate exported `Group`, `Version`, and `GroupVersion` constants for
each generated kind's package (when grouping kinds by version, only one
set of constants will be generated for the package). This allows for a
more convenient way to reference these constants vs. using the Kind
accessor method, and then getting its Group/Version, especially when
kinds are grouped by version, so multiple kinds are in the same package,
sharing the same constants.

This is implemented as a new generated file because it has to be
generated only once per package, unlike the other kind files which exist
once per kind.

Resolves https://github.com/grafana/grafana-app-sdk/issues/515