[logging] Replace logging with local dir (for now)
We have a chicken and egg problem: we can't push the logging submodule
without depending on it, and we can't depend on it without pushing it,
making the build fail.

To avoid this, we'll use a `replace` directive (for now?), such that the
local directory is used directly instead of fetching from Git.