Small fixes for the Issue Tracker tutorial (#619)
This PR updates the Issue Tracker tutorial with few small fixes.

Apart from these fixes, I had to change `issue-tracker-project-app` part
of URL (eg. in `curl
http://grafana.k3d.localhost:9999/api/plugins/issue-tracker-project-app/resources/v1/issues`
command) to `issuetrackerprojectapp`. I'm not sure why, or whether this
is specific to my installation somehow, or version, so I didn't update
the commands in this PR.