Add `--noschemasinmanifest` flag to CLI generate command (#623)
Add `--noschemasinmanifest` flag to CLI generate command, which excludes
the kind schemas when generating the app manifest. This is required for
schemas which are recursive that cog permits codegen for, but CUE's
openapi generation, used for creating CRD schemas, does not.

This is a temporary fix until a better resolution can be made by
resolving https://github.com/grafana/grafana-app-sdk/issues/460