[OpinionatedWatcher] Address bug where deletes on add-in-progress or add-retry resources aren't propagated (#630)
Address bug where deletes on add-in-progress or add-retry resources
aren't propagated. Use an additional "wip" finalizer that gets attached
at the beginning of the add, before the downstream add call, and is then
replaced by the "real" finalizer on completion of the add. The updated
workflow for add is now:
1. Check if the resource is actually deleted (this is for operator
start-up where DeletionTimestamp != nil). Do delete logic and remove
finalizer(s) if so.
2. Check if the object has the "real" finalizer. If so, call sync
instead
3. Add the "wip" finalizer (if missing)
4. Call downstream Add function
5. On success, replace "wip" finalizer with "real" finalizer

On failure, "wip" finalizer stays for retries. If no retries happen, and
an update comes in for the object later, it will be forwarded back to
`Add`. If a delete comes in, downstream `Delete` will be called and the
finalizer removed on success.

Resolves https://github.com/grafana/grafana-app-sdk/issues/46