[AppManifest] Add tests for generated AppManifest code and conversion code (#658)
This is a follow-up to
https://github.com/grafana/grafana-app-sdk/pull/639

Add tests for generated AppManifest code and conversion code. As part of
this, I realized that some fields (`extraPermissions` and
`selectableFields`) should be optional in the CUE, so as not to be
required in the CRD, so I updated those fields and re-ran `make
generate`.