[codegen] Fix bug in CRD generation where 'additionalProperties' is not parsed in arrays (#659)
Resolves https://github.com/grafana/grafana-app-sdk/issues/645

Codegen for CRDs and schemas in the app manifest was not properly
parsing openAPI array types to replace `"additionalProperties": {}` with
`"x-kubernetes-preserve-unknown-fields": true`.

The code that did this replacement (`replaceAdditionalProperties`) was a
bit hard to follow, so I both fixed it there, and also introduced a new
`setKubernetesPreserveUnknownFields` method which I think is a bit
easier to follow and may be a bit more extensible. Please let me know if
this is harder to follow, or if the previous, less-lines-of-code method
feels like the better solution to continue with.

Added another test for this into the test CUE, and regenerated golden
test files.