Add typed and namespaced resource clients
What

This change adds two new resource clients - a typed client, which uses
generics and ensures the GVK is always properly set on mutations
and a namespaced client, which builds on top of the typed client and
additionally sets the namespace to all requests, based on the value
passed in initializer.

Why

These clients are more convenient to work with, when a store is not
required, e.g. in short-lived processes like CLIs or other tools which
interact with the APIs.

We could e.g. use these in codegen for providing instantiated clients for generated kinds.

Closes https://github.com/grafana/grafana-app-sdk/issues/690

Signed-off-by: Igor Suleymanov <igor.suleymanov@grafana.com>