[codegen] Don't use `prometheus.MustRegister` in generated plugin code's `newInstanceFactory` (#713)
Don't use `prometheus.MustRegister` in generated plugin code's
`newInstanceFactory `to avoid panicking if and when it is called again
during the plugin lifetime.

In order to still denote an error if a non-already-registered error
occurs, check if the error is not an instance of
`prometheus.AlreadyRegisteredError`, and return an error from the
factory method if so.