Add operator health endpoints and share mux with metrics (#626)
### Introduction

Adds functionality that allows app developers to specify their own
implementation of health checks. The inspiration for this came from
`metrics.Provider`, which is an interface. An informer, watcher or
controller is able to implement this interface and the runners
automatically expose metrics from the said components. Similarly, with
this PR, we introduce two Interfaces in the brand new `health` pacakge
that app developers can optionally implement their own checks using the
`Checker` interface, each checks implementation being an implementation
of `Check` interface.

Additionally, the `health` package provides the `Observer`
implementation. Observer runs the aforementioned checks periodically and
a cached value can be fetched at any time. This is to prevent a DDoS
attack in case an external actor exercises the status endpoints
ad-infinitum.

#### Details

* The checks implementation is merged with the server mux that already
exists in the metrics server. Some newer config structs are introduced
to streamline providing the port and interval configuration. `9090`
being the default for backwards compat when such config is omitted.
* `operator/runner` and `app/runner` are updated to provide this
functionality.
* `simple` package (included the deprecated operator) components are
also updated.
* `examples` were minimally updated as well to pass in the new config.