Fix resource.CopyObject and resource.CopyObjectInto zeroing time.Time (#744)
`resource.CopyObject` and `resource.CopyObjectInto` incorrectly zero
`time.Time` values when doing their copy, as it tries to copy
`time.Time` instances as structs using reflection. Since `time.Time` has
no exported fields, this does not work. To fix this behavior, check if
the struct to be copied is `typeOf(time.Time)`, and, if so, directly
copy it with `dst.Set(src)`, rather than recursively calling
`CopyObjectInto` to copy it as a user-defined struct.

This does bring up the issue that any other "special" types which have
custom `MarshalJSON`/`UnmarshalJSON` implementations but no exported
fields will also fail with the deep copy performed by
`resource.CopyObject`/`resource.CopyObjectInto`. I'm not sure if this is
a restriction we want to live with, or attempt to solve here.

Ultimately, a better solution would be to avoid using relfection for
copying objects entirely, and rely on generated deep copy methods in
objects (see https://github.com/grafana/grafana-app-sdk/pull/424).