Fix resource.CopyObject for *time.Time elements. (#747)
Previous PR to fix `resource.CopyObject` and `resource.CopyObjectInto`
for `time.Time` elements
(https://github.com/grafana/grafana-app-sdk/pull/744) had a small bug
with `*time.Time` types, and the addition to the test only tested
`time.Time`. Added a `*time.Time` element to the test, and fixed the bug
where it improperly checked if the pointer is of type `time.Time`,
rather than the dereferenced pointer.