operator: Add initial concurrent reconciliation support (#724)
### What

This PR adds support for reconciling events concurrently in
`operator.InformerController`. By default only `1` worker would be run,
to maintain backward compatibility of processing the events
sequentially.

This is done with the help of a new `ResourceWatcher` implementation
called `ConcurrentWatcher`, which runs multiple workers in the
background and shards the incoming events between them. For every event,
a hash mod is calculated on the identifier, allowing us to deliver
events for same object to the same worker. This is done to maintain the
per object in-order guarantee for events by Kubernetes.

### Why

This is needed for
https://github.com/grafana/cloud-apps-platform/issues/883

---------

Signed-off-by: Prem Saraswat <prmsrswt@gmail.com>