Fix Broken Retry Logic in Apps (#758)
Since
[v0.32.0](https://github.com/grafana/grafana-app-sdk/releases/tag/v0.32.0),
the `OpinionatedWatcher` uses an "add-in-progress" finalizer that gets
added at the beginning of an Add operation, and removed at the
conclusion of a successful add. Adding this finalizer triggers an
update, which is swallowed by the `OpinionatedReconciler` automagically.

If the downstream Add operation fails, the Add is added to the retry
queue. However, since events are serialized, right after the add failure
is put into the retry queue, the Update event for the finalizer addition
is received by the InformerController. If the InformerController has no
`RetryDequeuePolicy`, retries for an object are automatically dequeued
when a new event comes in, so the retry of the failing add is
immediately dequeued, and never retried. This is resolved by the
`OpinionatedRetryDequeuePolicy`, which will not dequeue events which are
different operations (Add and Update). However, by default,
`RetryDequeuePolicy` is nil on a new `InformerController`, and there is
a bug in `simple.NewApp` where the `RetryDequeuePolicy` (and,
incidentally, the `RetryPolicy` as well) is not copied into the
underlying `InformerController`, even if present in the config.

This PR does two things:
* Has the default `InformerControllerConfig` generated by
`operator.DefaultInformerControllerConfig()` use the
`OpinionatedRetryDequeuePolicy`. Both `simple.NewApp()` and
`simple.NewOperator()` use `operator.DefaultInformerControllerConfig()`
to construct their initial `InformerController`, so this ensures that
default apps will not encounter this retry dequeue bug.
* Correctly propagates the `RetryDequeuePolicy` (and `RetryPolicy`)
specified in `simple.AppInformerConfig` into the created
`InformerController` (if such policies are present).

This PR also adds additional context to the godoc for the
`InformerController`'s `RetryDequeuePolicy` field, which notes that a
policy should be present for using an `OpinionatedWatcher` in order for
retries of failed Add operations to succeed.

I plan to put together another PR to try to introduce testing which
would have caught this regression (and similar future ones), but it's a
bit trickier as it depends on apiserver behavior (the patch triggering
an update event in the watch stream), and we don't mock out an
apiserver.