[Local Dev Environment] Remove Anonymous Auth in local dev environmen… (#738)
Remove Anonymous Auth in local dev environment grafana instance, update
docs to reflect this change.

Related to https://github.com/grafana/grafana-app-sdk/pull/737, as the
anonymous role of `Admin` is being deprecated, and a `Viewer` anonymous
role doesn't show the plugin, so for ease-of-use anonymous auth will be
turned off. It can be re-enabled with the `Viewer` role via
`local/config.yaml` by adding
```
grafanaWithAnonymousAuth: true
```

We could also have `grafanaAnonymousAuth` be a string key that allows
for specifying a specific role, rather than an on/off switch, I have no
strong opinions on that, and would love any opinions on it.