[AppManifest] Add generated function to associate a kind and version to its go `resource.Kind` (#775)
To build an apiserver runner, the go types for an app must be registered
in the scheme prior to the point of the loopback interface creation to
provide as a kubeConfig to the `NewApp` function. This creates a slight
chicken-and-egg problem, where the app is the only thing that knows the
specific go types for the kinds it manages, but needs a kubeConfig to be
initialized for `ManagedKinds()` to be called by the runner. The
AppManifest data (`app.ManifestData`) has all the information on the
kinds managed by the app, _excepting_ the go types it uses, as the
manifest data is assumed to be file-like, and can be loaded from disk or
an apiserver as well.

This PR is a proposal for a solution to this problem, by adding a
generated function alongside the generated manifest that associates a
kind (string) and version (string) to the app's corresponding generated
`resource.Kind` implementation. The apiserver runner config can than
consume that function in its `New` method, and allow it to make the
associations based on the manifest for the app prior to loading the app
itself. If an app author does not use the codegen, they can simply
implement this function as well if using the apiserver runner.