[AppManifest] Implement `app.VersionSchema.AsKubeOpenAPI` (#779)
To move toward enabling apiserver functionality without needing the
`--postprocess` flag when generating code, we need to be able to provide
a function that consumes a kubernetes `kube-openapi`
`common.ReferenceCallback` and returns a
`map[string]common.OpenAPIDefinition`. Currently, this is generated by
using kubernetes `openapi-gen` in the jenny that runs for
`--postprocess`, but this sources from generated go code, and
information can be lost in the conversion from CUE->go->OpenAPI.
Additionally, we would like to be able to source everything required for
an app to run from the AppManifest itself, and not require extra
generated code for the runner.

This PR implements the function `AsKubeOpenAPI` on the AppManifest's
`VersionSchema` object, which produces similar output to the output
created by generated code. This allows an apiserver runner to require
only the manifest for getting the required kubernetes openAPI objects,
rather than needing to reference generated code.