Allow Subresource Patching in PatchOptions and k8s.Client (#796)
Add a `Subresource` field to `PatchOptions` similar in behavior to the
`Subresource` field in `UpdateOptions`, to allow a `Patch` request to
patch a subresource.

Noted in godoc that for an empty subresource, the entire subresource
must be set in the patch request. Otherwise, kubernetes responds with a
422 status and a message of "the server rejected our request due to an
error in our request" (with no other details).