Add local zizmor config (#805)
For PR's that are sourced from a fork, the zizmor check will fail
without a config present in the fork's `.github` directory. There is
work being done to fix this upstream in the reusable workflow itself,
but for now, we need the file in the local repo so it will be present in
the fork.