[Tutorial] Small fixes (#820)
Thank your for this tutorial, I though it was excellent.

While following it, I came up with a small set of nits.

I also run into the issue mentioned
[here](https://raintank-corp.slack.com/archives/C04RVCAG9B5/p1747394717867569),
which was fixed by the suggestion to use
`grafana/grafana-enterprise:12.0.0` as the grafana image in
`local/config.yaml`.

Finally, while working on admissions controls and adding the validation,
somehow I managed to really break the operator (and all interactions
with the API), by creating an invalid Issue before adding the
validation. However it is very possible it was my fault :tm: and that I
forgot to generate before the the rebuild. In any case, it was fixed
after recreating the cluster. And did not happen again.

Once again, thank you for this excellent resource.