[codegen] Support non-string selectable fields of types (#825)
Support non-string selectable fields of types `int` (including `int32`
and `int64`), `bool`, and `time.Time`. Field type determination is a bit
messy for now.

Previously, codegen created non-compiling go code for the
`_schema_gen.go` file for a kind if a selectable field type was not a
string. This behavior is supported in CRDs, so should also be supported
in the API server code created in `_schema_gen.go` for the field. This
PR adds a slightly-guesstimating-way of determining the type of each
field in the selectable field list, and then formatting that type to a
string in the resultant go code.

Resolves https://github.com/grafana/grafana-app-sdk/issues/809

---------

Co-authored-by: Costa Alexoglou <konsalexee@gmail.com>