API Server Installer (#811)
alternate to https://github.com/grafana/grafana-app-sdk/pull/792

@IfSentient Notes:
This PR adds the ability to run an API server for an app given an
`app.Provider` and `ManagedKindResolver` (function generated from CLI).
New code is in `k8s/apiserver` (with example in `examples/apiserver`),
and small bugfixes in `app/manifest.go` for `AsKubeOpenAPI` to function
correctly with kubernetes' reflection magic. A few new types have been
added, broken down here:
* `apiserver.AppInstaller` - this is an interface which describes all
the methods needed to "install" an app onto a typical API server. We
provide a default (unexported currently to prevent incorrect
instantiation) implementation of this interface with
`apiserver.NewAppInstaller`.
* `apiserver.Options` - extends the kubernetes
`options.RecommendedOptions` and used to create a Config. Takes a list
of `AppInstaller` instances on new.
* `apiserver.Config` - extends the kubernetes `server.RecommendedConfig`
and used to create a Server.
* other unexported types kubernetes requires implementations of: 
    * `appAdmission` - admission plugin used for validation and mutation
    * `strategy` - required for storage

~2600 lines of this PR is the copy-pasted code for the common API types
in `k8s/apiserver/openapi.go`.

The `apiserver.AppInstaller` is intended to be used in OSS grafana to
replace the app builder at some point in the future.

---------

Co-authored-by: Austin Pond <austin.pond@grafana.com>