Custom Subresource Routes (#852)
# What this PR Does

This PR adds the ability for the apiserver/default API server installer
to expose custom subresource routes. It uses the existing format for
specifying custom subresource routes in kinds in the AppManifest to
expose them in the apiserver itself. This requires some additional types
to be supplied by the app author, so additional codegen has been added
to generate go types from the custom route responses, and a new type
associator function (like the one for kinds) is also now generated as
part of the manifest go file.

There is also a drive-by fix in the `k8s.DynamicPatcher` for handling
behavior noticed in with the apiserver kinds' reconcilers.