API Server & Custom Subresource Route Support (#864)
This PR introduces a major change to the AppManifest, including the CUE
manifest and kind definitions for codegen, as well as the
`k8s/apiserver` package which allows an app author to build an API
server for an app, or install an app into an existing API server. The
API server support includes support for custom subresource routes, and
relies on the data in the AppManifest (so a complete manifest must
exist, without `--noschemasinmanifest` being used for generation).

All changes except a final `main` merge in this PR have been previously
reviewed (the `feat/apiserver` branch requires PR's), see those PR's for
full contexts on each change:
* [AppManifest & Codegen
Changes](https://github.com/grafana/grafana-app-sdk/pull/795)
* [API Server](https://github.com/grafana/grafana-app-sdk/pull/811)
* [Custom Subresource Route Support in the API
Server](https://github.com/grafana/grafana-app-sdk/pull/852)
* [Merge main into
feat/apiserver](https://github.com/grafana/grafana-app-sdk/pull/858)
* [Documentation
Updates](https://github.com/grafana/grafana-app-sdk/pull/863)

The final `main` merge done after the previous PR's can be seen in the
commit list or directly viewed [here
(4a87157)](https://github.com/grafana/grafana-app-sdk/pull/864/commits/4a871571376c44de3f37874a4789372eb5b5165c)

## AppManifest & Codegen Changes

Full context in [this
PR](https://github.com/grafana/grafana-app-sdk/pull/795)

The AppManifest has been changed to contain `versions` which contain
`kinds`, instead of `kinds` which contain `versions`. The reasoning for
this change is that in an API server, everything it determined on a
GroupVersion level, and adding custom route support needs to be at a
version level across the whole group, not per-kind. This also forces an
app author to think in terms of versions being the top-level construct,
rather than kinds, which is how the API server behaves.

This results in slightly more complex CUE, as each versioned kind needs
to contain redundant information that matches exactly the kind of the
same name in a different version, but we have adopted a standard pattern
in the generated manifest and kind CUE to make this simpler.

manifests/kinds can still be parsed the "old" way with the CLI using the
`--useoldmanifestkinds` flag. The produced AppManifest will be written
the "new" way.

## API Server Support

Full context in [this PR to add the API
server](https://github.com/grafana/grafana-app-sdk/pull/811), and [this
PR to enable custom subresource
routes](https://github.com/grafana/grafana-app-sdk/pull/852)

The `k8s/apiserver` package has been introduced to allow an app author
to run their app as a standalone API server, or to use the
`apiserver.Installer` to install their app onto an existing API server
(such as the grafana API server). As part of this, custom subresource
routes, as defined in the manifest and implemented in the app, are
supported, along with all existing behavior that is supported for
operator apps.

An example API server app has been placed in `examples/apiserver`.

## Documentation Updates

Docs have been updated to reflect the changes in the
AppManifest/codegen. No documentation aside from godocs and the example
exist yet for the API server code, they will be introduced in a future
PR.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Signed-off-by: Igor Suleymanov <igor.suleymanov@grafana.com>
Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>
Co-authored-by: Todd Treece <360020+toddtreece@users.noreply.github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Stephanie Hingtgen <hingtgen.stephanie@gmail.com>
Co-authored-by: Jorge Creixell <jcreixell@gmail.com>
Co-authored-by: Oleg Zaytsev <mail@olegzaytsev.com>
Co-authored-by: Costa Alexoglou <konsalexee@gmail.com>
Co-authored-by: Costa Alexoglou <konstantinos.alexoglou@grafana.com>
Co-authored-by: Daniele Stefano Ferru <daniele.ferru@grafana.com>
Co-authored-by: Igor Suleymanov <radiohead@users.noreply.github.com>