[k8s.apiserver] Fix the ParameterCodec used by InstallAPIs  (#886)
Fix the ParameterCodec used by InstallAPIs to use the installer scheme
rather than the metav1 scheme's ParameterCodec.

This was resulting in a bug in certain situations with `CallResource`
where an error would be returned that the kind couldn't be found in the
scheme, this was due to the decoding using the ParameterCodec from the
`metav1` package, which is based on the unexported metav1 scheme.