Add Custom Subresource Request Method to Clients (#879)
Now that arbitrary custom subresources are supported for API server
apps, we need to update the `resource.Client` (and other clients built
on top of it) to allow for making arbitrary subresource requests. This
PR introduces a new method:
```go
SubresourceRequest(context.Context, resource.Identifier, resource.CustomRouteRequestOptions) ([]byte, error)
```
to `resource.Client` (and similar methods to `resource.TypedClient` and
`resource.NamespacedClient`), and implements it in `k8s.Client`. The
underlying implementation in `k8s.gvclient` can also be used for custom
resource routes once implemented.

Another improvement to be made in a subsequent PR will be to generate go
clients for each kind that include dedicated, strongly-typed methods for
handling each custom subresource.

Resolves https://github.com/grafana/grafana-app-sdk/issues/877