Use cog for CRD/AppManifest OpenAPI Generation (#876)
This PR switches to using [grafana/cog](https://github.com/grafana/cog)
instead of the CUE SDK for OpenAPI generation to use in CRDs and the
AppManifest. This both unifies our code generation tooling to only using
cog, as well as solves the problem with recursive CUE definitions
causing the `generate` command to hang and eventually crash due to the
CUE SDK's OpenAPI generation recursing indefinitely (cog's intermediate
AST allows for the OpenAPI generation to work properly with recursion).

The OpenAPI returned by cog contains a list of definitions, so this is
condensed into a single CRD-compatible OpenAPI schema, with recursive
elements turned into objects with
`x-kubernetes-preserve-additional-properties` set to `true`.

As we now support more than CRDs with the OpenAPI, the next step after
this PR is to move the App Manifest's version's kind's schemas to use a
full OpenAPI components set like what is generated by cog (which can
properly represent the recursion by using definition references, which
are unsupported by CRDs), and move the
components-to-single-CRD-compatible-object logic into the AppManifest to
allow the schema to be fetched as either a CRD-compatible schema or one
fit to use in an API server or other machinery which can handle
definition references. This was left out of this PR as this PR is
designed to keep the current behavior while switching the OpenAPI
generator.

Resolves https://github.com/grafana/grafana-app-sdk/issues/460

---------

Co-authored-by: Igor Suleymanov <radiohead@users.noreply.github.com>